requirements.txt
PyYAML==6.0
torch==1.10.2
torchaudio==0.10.2
librosa==0.10.0.post2
tqdm==4.65.0
pytorch-lightning==1.5.1
urllib3==1.26.15
transformers==4.28.1
dtw==1.4.0
python_speech_features==0.6
hmmlearn==0.3.0
typeguard==4.0.0
overrides==7.3.1
jsonnet-binary==0.17.0
setuptools==59.5.0
pytorch-pretrained-bert==0.6.2
webrtcvad==2.0.10
wave==0.0.2
